scrape_data/requirements.txt
beautifulsoup4
psycopg2-binary
requests
selenium
webdriver-manager
